pyproject.toml
[project]
name = "slack-rection-to"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "fastapi>=0.115.12",
    "google-cloud>=0.34.0",
    "pydantic>=2.11.5",
    "pydantic-ai>=0.2.7",
    "pyyaml>=6.0.2",
    "ruff>=0.11.11",
    "slack-bolt>=1.23.0",
    "slack-sdk>=3.35.0",
    "uvicorn>=0.34.2",
]
